register custom converter for instant to offsetdatetime for projections
spring data jpa cannot convert from instants to offsetdatetime automatically in projections. For full entities this is not an issue as hibernate seems to take over the job, however the projections sometimes run into errors because the returned field instance is an instant which cannot be converted to an offsetdatetime. Second hurdle is that a custom converter cannot be registered easily, there is a github issue about this in spring but without an answer yet. Therefore, the current code just relies on a hack to register the converter.