[ci skip] schedule ci action execution once per day
This is due to constant changes in the SimRail backend data, e.g. new points. Without having the action scheduled, there is no real way to find out if some data is missing in the backend files without having luck and finding it by accident when running the tests manually