reset foreign id when a journey is removed
It can happen that a train is marked as removed but respawns later, for example if it derails right before a server restart close to the starting point. In that case the respawn is not properly recorded anymore due to the fact that the foreign id is already set. With this patch the foreign id of the journey is reset to ensure that a respawn gets recorded if one occurs. The last seen time gets removed on spawn too, as the journey wasn't really seen for the last time if it respawns.