feat: expose server utc offset hours, better integrate the time data
The offset hours are already known to the data collector to provide the realtime time information when a train arrives at a station. However, the offset is also useful for external developers to display the correct server date and for the api endpoints to fill missing time and date values (which currently default to the utc time) with the actual server time/date.