feat: new event subscription system (#43)
The old event system was quite outdated on the way how event updates to
receive were selected. The events could only be filtered by server and
had to be provided during the initial connection in the query string.
The new event system addresses this problem in the following ways:
1. For subscriptions a text message must now be sent to the client,
allowing to change subscriptions without reconnecting.
2. Data can now be filtered by their respective journey/dispatch post id
as well, allowing to only subscribe to updates of a single journey
rather than all journeys on a server
3. Subscriptions can now be removed again without needing to reconnect
by sending a text message
4. The whole event system is now documented for other developers to
easily understand how to access data